Create rc_unstability_explanation.py
Hypothesis: To be a RC member, you have to be within a given range of chasing rank (distribution). But Having a particular social rank does no guarantee being in RC. To test this, generate random draws from RC rank distribution and see how often they coincide. This model agrees to 0.01% difference with experiement!!!