Critical bug in the order (or ranking) by chasing and approaches.
I was using the np.argsort function, which does not return the order but the indices that would be necessary to sort. So to get the order of an array, one needs to use list(array).index(idx)!! This drastically changed the results for the approach order of the rich club.